requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.3.1
annotated-types==0.7.0
anyio==3.6.2
appdirs==1.4.4
APScheduler==3.11.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.8.1
asttokens==2.2.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==23.2.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.0
billiard==4.2.1
bleach==6.0.0
blinker==1.9.0
blis==0.7.10
boilerpy3==1.0.7
breadability==0.1.20
bs4==0.0.1
cachetools==5.5.2
catalogue==2.0.9
cattrs==23.2.3
celery==5.4.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.1.0
chord==6.0.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.1.2
comtypes==1.4.8
confection==0.1.3
contourpy==1.0.7
crispy-bootstrap4==2024.10
cron-descriptor==1.4.5
cycler==0.11.0
cymem==2.0.8
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dj-database-url==2.3.0
Django==5.1.6
django-apscheduler==0.7.0
django-celery-beat==2.7.0
django-crispy-forms==2.3
django-timezone-field==7.1
docopt==0.6.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
Events==0.5
executing==1.2.0
farm-haystack==1.26.2
fastjsonschema==2.16.3
filelock==3.15.4
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
funcy==2.0
gast==0.6.0
gensim==4.3.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
google-pasta==0.2.0
googleapis-common-protos==1.69.0
grpcio==1.71.0rc2
grpcio-status==1.71.0rc2
gunicorn==23.0.0
h11==0.14.0
h5py==3.11.0
haystack-ai==2.3.0
haystack-experimental==0.1.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.5
idna==3.4
importlib-metadata==6.6.0
inflect==7.3.1
inflection==0.5.1
ipykernel==6.22.0
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
jsonpickle==3.0.4
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.6
keras==3.8.0
kiwisolver==1.4.4
kombu==5.4.2
langcodes==3.3.0
lazy-imports==0.3.1
libclang==18.1.1
lxml==4.9.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
ml-dtypes==0.4.0
monotonic==1.6
more-itertools==10.3.0
multidict==6.0.5
murmurhash==1.0.10
namex==0.0.8
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.3
nltk==3.8.1
notebook==6.5.3
notebook_shim==0.2.2
num2words==0.5.13
numexpr==2.10.0
numpy==2.0.2
openai==0.28.0
opencv-python==4.7.0.72
opt-einsum==3.3.0
optree==0.12.1
packaging==23.0
pandas==2.0.0
pandoc==2.3
pandocfilters==1.5.0
parso==0.8.3
pathy==0.10.2
pickleshare==0.7.5
pillow==11.1.0
platformdirs==3.1.1
plumbum==1.8.1
ply==3.11
posthog==3.5.0
preshed==3.0.9
prometheus-client==0.16.0
prompt-toolkit==3.0.38
prompthub-py==4.0.0
proto-plus==1.26.0
protobuf==5.29.3
psutil==5.9.4
psycopg2-binary==2.9.10
pure-eval==0.2.2
pyairtable==3.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==23.12.11
pycparser==2.21
pydantic==2.10.6
pydantic_core==2.27.2
pyee==8.2.2
Pygments==2.14.0
pyjokes==0.8.3
pyLDAvis==3.4.1
pyparsing==3.0.9
pypiwin32==223
pyppeteer==1.0.2
pyrsistent==0.19.3
python-crontab==3.2.0
python-dateutil==2.8.2
python-decouple==3.8
python-json-logger==2.0.7
pyttsx3==2.98
pytz==2023.3
pyvis==0.3.2
PyWavelets==1.4.1
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.1
QtPy==2.3.0
quantulum3==0.9.2
rank-bm25==0.2.2
redis==5.2.1
regex==2023.8.8
requests==2.30.0
requests-cache==0.9.8
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.12.2
Send2Trash==1.8.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.4
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.4.7
sseclient-py==1.8.0
stack-data==0.6.2
sumy==0.11.0
tenacity==8.5.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.4.0
terminado==0.17.1
thinc==8.1.12
threadpoolctl==3.4.0
tiktoken==0.7.0
tinycss2==1.2.1
tokenizers==0.15.2
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.39.3
typeguard==4.3.0
typer==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.3.1
uri-template==1.2.0
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==1.26.15
vine==5.1.0
wasabi==1.1.2
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
Werkzeug==3.1.3
whitenoise==6.9.0
widgetsnbextension==4.0.6
wordcloud==1.9.3
wrapt==1.16.0
yarl==1.9.4
zipp==3.15.0
